requirements.txt
# モダンGUI・テーマ関連（新しいUI）
customtkinter>=5.2.0

# 旧UIとの互換性（既存コードで使用）
ttkthemes
ttkbootstrap>=1.19.0

# 画像処理 (ファビコン・プレビュー表示用)
Pillow>=9.0.0

# ネットワーク・スクレイピング (プレビュー取得用)
requests>=2.28.0
beautifulsoup4>=4.11.0

# AI関連 (Gemini API)
google-generativeai>=0.3.0

# ユーティリティ
pyperclip>=1.8.2